requirements.txt
annotated-types==0.7.0
dnspython==2.6.1
load-dotenv==0.1.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dotenv==1.0.1
typing_extensions==4.12.2
